requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
dnspython>=2.4.0
paramiko>=3.3.0
colorama>=0.4.6